Improved join cardinality estimation, replacing PropertySelectivityCardinalityHeuristic with GeneralSelectivityHeuristic.
Join cardinality estimation was also changed so that estimation is done by generating a new CQuery that represents the joined nodes and then estimating it with the injected CardinalityEnsemble, which in the default configuration has the new heurustic.

The estimator applies heuristics for beneffiting star queries, and for "smearing" the selectivity of triples to one another.