SPARQLClient uses httpcomponents-client to stream-parse TSV results from SPARQL endpoints.
In comparison to ARQ on Virtuoso endpoints, total consumption got a speedup of ~1.66 times while speedup for first result is ~3.70 times. The drawback is that Virtuoso does not respect the standard RDf serialization for TSV and uses CSV serialization instead, which looses RDF node type information (no datatypes and URIs/literal ambiguity).

Fixinf SPARQLClient parsing and integrating it into unit  and integration tests revealed bugs in the parsing of the results from the LargeRDFBench. Test data was updated to restore discarded literal datatypes.