Changed StdPrefixDict implementation from a trie to a custom IndexedSet-based implementation. SPARQLParser copies the jena PrefixMapping to a StdPrefixDict.
The new implementation is 6.5-7.5 times faster on shortening URIs than tries, which where already faster than Jena's PrefixMapping. For iterating the prefixes, it is 0.9~1.2 times faster than jena, which was faster than the trie implementation. For building, it is 1.17 times slower for small maps, but is faster for larger maps (using .builder()).